[CON-96] Allow pronouns in answer
