+++ lib/data/utils/pronoun_extensions.dart
@@ -1,0 +1,21 @@
+import '../../utilities/extensions/string_extensions.dart';
+import '../enums/pronoun.dart';
+
+extension PronounExtensions on Pronoun {
+  bool _canBeGendered(String italianParticiple) {
+    if (isPlural) {
+      return ['i', 'e'].contains(italianParticiple.last);
+    } else {
+      return ['o', 'a'].contains(italianParticiple.last);
+    }
+  }
+
+  String genderItalianConjugationIfPossible(String italianParticiple, {bool forceGender = false}) {
+    if (forceGender || _canBeGendered(italianParticiple)) {
+      return isPlural ?
+      italianParticiple.replaceLastWith('i/e') : italianParticiple.replaceLastWith('o/a');
+    } else {
+      return italianParticiple;
+    }
+  }
+}
